agent-foundry/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-foundry"
version = "0.1.0"
description = "The enterprise agent incubation platform — from idea to governed production agent"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Ravi Labs", email = "ravi@ravi-labs.dev" },
]
keywords = [
    "ai", "agents", "policy", "governance", "incubation",
    "financial-services", "erisa", "compliance", "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
]
dependencies = [
    "pyyaml>=6.0.0",
    "click>=8.0.0",
]

[project.scripts]
foundry = "foundry.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
redis = [
    "redis[hiredis]>=5.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
encryption = [
    "cryptography>=41.0.0",
]
enterprise = [
    "agent-foundry[redis,otel,encryption]",
]

[project.urls]
Homepage = "https://github.com/ravi-labs/agent-foundry"
Repository = "https://github.com/ravi-labs/agent-foundry.git"
Documentation = "https://github.com/ravi-labs/agent-foundry#readme"
Issues = "https://github.com/ravi-labs/agent-foundry/issues"
"Tollgate (Policy Engine)" = "https://github.com/ravi-labs/tollgate"

[tool.hatch.build.targets.wheel]
packages = ["src/foundry"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PTH", "RET"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG001", "ARG002", "E402", "E501", "B007"]
"examples/**" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
